move lock service from utils
