+++ web-app/src/lib/services/lock.service.ts
@@ -1,0 +1,74 @@
+import { getEnvSecrets } from "@/config/env.config";
+import { prisma } from "@/lib/db";
+
+export async function getLock(lockKey: string) {
+  return await prisma.$transaction(async (tx) => {
+    // Example: Try to acquire a lock on a specific agent
+    // Using pessimistic locking with 'FOR UPDATE'
+    const lock = await tx.lock.findFirst({
+      where: {
+        key: lockKey,
+      },
+    });
+
+    if (!lock) {
+      return await tx.lock.create({
+        data: {
+          key: lockKey,
+          lockedBy: getEnvSecrets().INSTANCE_ID,
+          lockedAt: new Date(),
+          isLocked: true,
+        },
+      });
+    }
+
+    if (lock.isLocked) {
+      if (
+        lock.lockedAt &&
+        lock.lockedAt < new Date(Date.now() - getEnvSecrets().LOCK_TIMEOUT)
+      ) {
+        //TODO: better logging
+        console.warn(
+          "Lock timeout reached, will release key",
+          lockKey,
+          "last updated at: ",
+          lock.updatedAt,
+          " by instance: ",
+          lock.lockedBy,
+        );
+        return await tx.lock.update({
+          where: { id: lock.id },
+          data: {
+            lockedBy: getEnvSecrets().INSTANCE_ID,
+            lockedAt: new Date(),
+            isLocked: true,
+          },
+        });
+      }
+      return null;
+    }
+    return await tx.lock.update({
+      where: { id: lock.id },
+      data: {
+        lockedBy: getEnvSecrets().INSTANCE_ID,
+        lockedAt: new Date(),
+        isLocked: true,
+      },
+    });
+  });
+}
+
+export async function releaseLock(lock: { updatedAt: Date; key: string }) {
+  const updatedLock = await prisma.lock.update({
+    where: { key: lock.key, updatedAt: lock.updatedAt },
+    data: { isLocked: false, lockedBy: null, lockedAt: null },
+  });
+  if (!updatedLock) {
+    console.error(
+      "Lock changed while locked, will not release. Expected key",
+      lock.key,
+      "to be last updated at: ",
+      lock.updatedAt,
+    );
+  }
+}
